+++ src/main/kotlin/com/dodo/dodolistserver/auth/config/AuthConfig.kt
@@ -1,7 +1,7 @@
 package com.dodo.dodolistserver.auth.config
 
-import com.dodo.dodolistserver.auth.JWTTokenProvider
-import com.dodo.dodolistserver.auth.TokenProvider
+import com.dodo.dodolistserver.auth.provider.JWTTokenProvider
+import com.dodo.dodolistserver.auth.provider.TokenProvider
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
